requirements.txt
openai==0.27.8
tiktoken
langchain
langchainhub
python-dotenv
chromadb
bs4
html2text
sqlite3>=3.35.0